+++ rs/pocket_ic_server/src/pocket_ic.rs
@@ -22,7 +22,6 @@ use ic_config::{
     subnet_config::SubnetConfig,
 };
 use ic_crypto_sha2::Sha256;
-use ic_error_types::RejectCode;
 use ic_http_endpoints_public::{
     call_v2, call_v3, metrics::HttpHandlerMetrics, CanisterReadStateServiceBuilder,
     IngressValidatorBuilder, QueryServiceBuilder, SubnetReadStateServiceBuilder,
@@ -80,6 +79,7 @@ use pocket_ic::common::rest::{
     RawCanisterCall, RawCanisterId, RawEffectivePrincipal, RawMessageId, RawSetStableMemory,
     SubnetInstructionConfig, SubnetKind, SubnetSpec, Topology,
 };
+use pocket_ic::{ErrorCode, RejectCode, RejectResponse};
 use serde::{Deserialize, Serialize};
 use slog::Level;
 use std::hash::Hash;
@@ -121,6 +121,33 @@ pub(crate) type ApiResponse = BoxFuture<'static, (u16, BTreeMap<String, Vec<u8>>
 /// We assume that the maximum number of subnets on the mainnet is 1024.
 /// Used for generating canister ID ranges that do not appear on mainnet.
 pub const MAXIMUM_NUMBER_OF_SUBNETS_ON_MAINNET: u64 = 1024;
+
+fn wasm_result_to_canister_result(
+    res: ic_state_machine_tests::WasmResult,
+    certified: bool,
+) -> Result<Vec<u8>, RejectResponse> {
+    match res {
+        ic_state_machine_tests::WasmResult::Reply(data) => Ok(data),
+        ic_state_machine_tests::WasmResult::Reject(reject_message) => Err(RejectResponse {
+            reject_code: RejectCode::CanisterReject,
+            reject_message,
+            error_code: ErrorCode::CanisterRejectedMessage,
+            certified,
+        }),
+    }
+}
+
+fn user_error_to_reject_response(
+    err: ic_error_types::UserError,
+    certified: bool,
+) -> RejectResponse {
+    RejectResponse {
+        reject_code: RejectCode::try_from(err.reject_code() as u64).unwrap(),
+        reject_message: err.description().to_string(),
+        error_code: ErrorCode::try_from(err.code() as u64).unwrap(),
+        certified,
+    }
+}
 
 async fn into_api_response(resp: AxumResponse) -> (u16, BTreeMap<String, Vec<u8>>, Vec<u8>) {
     (
@@ -384,6 +411,9 @@ impl SubnetsImpl {
     pub(crate) fn get_all(&self) -> Vec<Arc<Subnet>> {
         self.subnets.read().unwrap().values().cloned().collect()
     }
+    fn clear(&self) {
+        self.subnets.write().unwrap().clear();
+    }
 }
 
 impl Subnets for SubnetsImpl {
@@ -421,8 +451,8 @@ pub struct PocketIc {
 
 impl Drop for PocketIc {
     fn drop(&mut self) {
-        let subnets = self.subnets.get_all();
         if let Some(ref state_dir) = self.state_dir {
+            let subnets = self.subnets.get_all();
             for subnet in &subnets {
                 subnet.state_machine.checkpointed_tick();
             }
@@ -452,8 +482,35 @@ impl Drop for PocketIc {
             let topology_json = serde_json::to_string(&raw_topology).unwrap();
             topology_file.write_all(topology_json.as_bytes()).unwrap();
         }
-        for subnet in subnets {
+        for subnet in self.subnets.get_all() {
             subnet.state_machine.drop_payload_builder();
+        }
+        let state_machines: Vec<_> = self
+            .subnets
+            .get_all()
+            .into_iter()
+            .map(|subnet| subnet.state_machine.clone())
+            .collect();
+        self.subnets.clear();
+        // for every StateMachine, wait until nobody else has an Arc to that StateMachine
+        // and then drop that StateMachine
+        let start = std::time::Instant::now();
+        for state_machine in state_machines {
+            let mut state_machine = Some(state_machine);
+            while state_machine.is_some() {
+                match Arc::try_unwrap(state_machine.take().unwrap()) {
+                    Ok(sm) => {
+                        sm.drop();
+                        break;
+                    }
+                    Err(sm) => {
+                        state_machine = Some(sm);
+                    }
+                }
+                if start.elapsed() > std::time::Duration::from_secs(5 * 60) {
+                    panic!("Timed out while dropping PocketIC.");
+                }
+            }
         }
     }
 }
@@ -1359,7 +1416,7 @@ fn process_mock_canister_https_response(
         reject_codes.push(reject_code)
     }
     for reject_code in reject_codes {
-        if RejectCode::try_from(reject_code).is_err() {
+        if ic_error_types::RejectCode::try_from(reject_code).is_err() {
             return OpOut::Error(PocketIcError::InvalidRejectCode(reject_code));
         }
     }
@@ -1431,7 +1488,7 @@ fn process_mock_canister_https_response(
         }
         CanisterHttpResponse::CanisterHttpReject(reject) => {
             CanisterHttpResponseContent::Reject(CanisterHttpReject {
-                reject_code: RejectCode::try_from(reject.reject_code).unwrap(),
+                reject_code: ic_error_types::RejectCode::try_from(reject.reject_code).unwrap(),
                 message: reject.message.clone(),
             })
         }
@@ -1553,7 +1610,7 @@ impl Operation for SubmitIngressMessage {
                     }
                     Err(SubmitIngressError::UserError(e)) => {
                         eprintln!("Failed to submit ingress message: {:?}", e);
-                        Err::<ic_state_machine_tests::WasmResult, ic_state_machine_tests::UserError>(e).into()
+                        OpOut::CanisterResult(Err(user_error_to_reject_response(e, false)))
                     }
                     Ok(msg_id) => OpOut::MessageId((
                         EffectivePrincipal::SubnetId(subnet.get_subnet_id()),
@@ -1616,16 +1673,18 @@ impl Operation for AwaitIngressMessage {
                         IngressStatus::Known {
                             state: IngressState::Completed(result),
                             ..
-                        } => return Ok(result).into(),
+                        } => {
+                            return OpOut::CanisterResult(wasm_result_to_canister_result(
+                                result, true,
+                            ));
+                        }
                         IngressStatus::Known {
                             state: IngressState::Failed(error),
                             ..
                         } => {
-                            return Err::<
-                                ic_state_machine_tests::WasmResult,
-                                ic_state_machine_tests::UserError,
-                            >(error)
-                            .into()
+                            return OpOut::CanisterResult(Err(user_error_to_reject_response(
+                                error, true,
+                            )));
                         }
                         _ => {}
                     }
@@ -1672,11 +1731,11 @@ impl Operation for IngressMessageStatus {
                     IngressStatus::Known {
                         state: IngressState::Completed(result),
                         ..
-                    } => Ok(result).into(),
+                    } => OpOut::CanisterResult(wasm_result_to_canister_result(result, true)),
                     IngressStatus::Known {
                         state: IngressState::Failed(error),
                         ..
-                    } => Err(error).into(),
+                    } => OpOut::CanisterResult(Err(user_error_to_reject_response(error, true))),
                     _ => OpOut::NoOutput,
                 }
             }
@@ -1703,15 +1762,20 @@ impl Operation for Query {
         match subnet {
             Ok(subnet) => {
                 let delegation = pic.get_nns_delegation_for_subnet(subnet.get_subnet_id());
-                subnet
-                    .query_as_with_delegation(
-                        self.0.sender,
-                        self.0.canister_id,
-                        self.0.method.clone(),
-                        self.0.payload.clone(),
-                        delegation,
-                    )
-                    .into()
+                match subnet.query_as_with_delegation(
+                    self.0.sender,
+                    self.0.canister_id,
+                    self.0.method.clone(),
+                    self.0.payload.clone(),
+                    delegation,
+                ) {
+                    Ok(result) => {
+                        OpOut::CanisterResult(wasm_result_to_canister_result(result, false))
+                    }
+                    Err(user_error) => {
+                        OpOut::CanisterResult(Err(user_error_to_reject_response(user_error, false)))
+                    }
+                }
             }
             Err(e) => OpOut::Error(PocketIcError::BadIngressMessage(e)),
         }