+++ rs/starter/src/main.rs
@@ -205,7 +205,7 @@ fn main() -> Result<()> {
         .arg("--config-file")
         .args([config_path.to_str().unwrap()]);
     info!(log, "Executing {:?}", cmd);
-    cmd.exec();
+    let _ = cmd.exec();
 
     Ok(())
 }