+++ rs/tests/consensus/orchestrator/rotate_ecdsa_idkg_key_test.rs
@@ -242,7 +242,7 @@ fn test(env: TestEnv) {
 
     assert!(last_rotation
         .duration_since(first_rotation)
-        .map_or(false, |d| d + gamma <= delta));
+        .is_ok_and(|d| d + gamma <= delta));
 
     // Ensure signing still works
     for (key_id, public_key) in public_keys {